Adds support for relative paths.
We have replaced the hard-coded absolute paths in various python and bash
scripts with relative paths.

Scripts automatically compute their own path and apply it to call other
scripts connected to them.

We have also distinguished between paths in the TG and paths in the SUT,
removing the assumption that in the TG and in the SUT the pastrami files
must be in the same path.

Signed-off-by: Paolo Lungaroni <paolo.lungaroni@uniroma2.it>